backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cohere>=5.20.0",
    "fastapi>=0.124.4",
    "openai>=2.13.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "uvicorn>=0.38.0",
]
